+++ src/otk/transform/__init__.py
@@ -4,7 +4,7 @@ import logging
 
 from typing import Any, Type
 
-from .context import Context
+from ..context import Context
 from .directive import op, define, include, desugar, customization
